Previously applications/configurations used a `set` instead of a `collection`
According to wiki an application id is `required` for configurations: https://github.com/lusis/Noah/wiki/Configuration

This change is mostly backwards compatible except a migration is may be needed for converting a `set` to a `collection`.

The old behaviour of creating a configuration without an application still exists. However it now uses `assert_unique` on `[:application_id, :name]`

Bumps the version to `0.9.0` so as not to break anyones application until ready.

Also added `em-hiredis`, `em-http-request` as dependencies for ease.

Added a test, all tests appear to be passing but please take a careful review for anything I may have missed.