Noah models - allow Sinatra to handle the exceptions.
Previous behavior was that lib/noah/models/* could return
exception as a string insted of an object. It's better to
let Sinatra to handle the exceptions.

This changes the exception returned by REST API, for example:
Old: undefined method `valid?' for #<String:0xDEADBEEF>
New: Connection refused - Unable to connect to Redis on ...